requirements.txt
cryptography==41.0.3
oci==2.90.3
Flask==2.1.0
PyYAML==6.0
requests==2.27.1
paramiko==3.3.1
Flask-Login==0.5.0
Flask-WTF==1.0.1
wtforms==3.0.1
Werkzeug==2.1.1
python-dotenv==1.0.0
